This is the June 2021 Update (v2.5.0) with production-ready OpenXR (#505)
* Add helper for making XrGraphicsBindingOpenGL* instances, start XR demo

* Fix OpenXR flags being 32-bit not 64-bit, "almost" working OpenXR demo

* Pushing latest demo - crosseyed is gone, replaced with weird fov thing

* Add maths helpers, is this progress I don't even know anymore

* Commit before rewrite

* Scrap the demo

* Stop us forgetting to ship bindings!

* Add missing bindings to SLN and missing typemaps, update changelog

* Update src/Core/Silk.NET.Core/Native/SilkMarshal.cs


* Bindings updates, fix release notes

* Update Silk.NET.sln

+++ src/OpenXR/Silk.NET.OpenXR/Enums/OverlaySessionCreateFlagsEXTX.gen.cs
@@ -11,7 +11,7 @@ namespace Silk.NET.OpenXR
 {
     [Flags()]
     [NativeName("Name", "XrOverlaySessionCreateFlagsEXTX")]
-    public enum OverlaySessionCreateFlagsEXTX : int
+    public enum OverlaySessionCreateFlagsEXTX : long
     {
     }
 }
